requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
mysql-connector-python==8.2.0
Werkzeug==3.0.1
